% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Loss function for the optimization problem}
\usage{
loss_function(
  tot_dur,
  accr_time = 24,
  h_c = 0.05/12,
  HR = 0.775,
  given_N = 10157
)
}
\value{
numeric
}
\description{
Loss function for the optimization problem
}
