% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_duration.R
\name{optim_duration}
\alias{optim_duration}
\title{Numerical optimization of total study duration}
\usage{
optim_duration(
  tot_dur,
  accr_time = 24,
  h_c = 0.05/12,
  HR = 0.775,
  uni_dim = TRUE
)
}
\value{
list of optimal duration and updated Nr. of required failures
}
\description{
Optimizes total study duration for a fixed accrual time. Hence, it returns
    optimal follow-up time after the last patient is enrolled for a fixed sample
    size. Thus, the function can be used to investigate follow-up variation by varying of
    baseline hazard, HR, or other quantities used for sample size calculation according
    to the exponential model given by Lachin et al (1986) (see package gsDesign).
}
