% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Loss function for the optimization problem}
\usage{
loss_function(
  tot_dur,
  accr_time = 24,
  h_c = 0.05/12,
  HR = 0.775,
  given_N = 10157
)
}
\arguments{
\item{tot_dur}{- scalar. Total study duration (accrual time + follow-up
time) in months.}

\item{accr_time}{- scalar. Accrual time in months.}

\item{h_c}{- scalar. Value of hazard rate in control arm.}

\item{HR}{- scalar. Hazard ratio.}

\item{given_N}{- integer. Fixed sample size.}
}
\value{
numeric
}
\description{
Loss function for the optimization problem
}
\details{
To avoid singular values a constraint "total time greater equal
    than accrual time" must be enforced.
}
